+++ examples/document-search/configurable.py
@@ -76,7 +76,7 @@ config = {
                 "score_threshold": -1.2,
             },
             "embedder": {
-                "type": "ragbits.core.embeddings.litellm:LiteLLMEmbedder",
+                "type": "ragbits.core.embeddings.dense:LiteLLMEmbedder",
             },
         },
     },